+++ packages/cloud_firestore/cloud_firestore/ios/Classes/FLTFirebaseFirestoreWriter.m
@@ -149,11 +149,15 @@
 }
 
 - (NSDictionary *)FIRDocumentSnapshot:(FIRDocumentSnapshot *)documentSnapshot {
-  FIRServerTimestampBehavior serverTimestampBehavior =
-      [self toServerTimestampBehavior:FLTFirebaseFirestorePlugin
-                                          .serverTimestampMap[@([documentSnapshot hash])]];
+  NSString *timestampBehaviorString =
+      FLTFirebaseFirestorePlugin.serverTimestampMap[@([documentSnapshot hash])];
 
-  [FLTFirebaseFirestorePlugin.serverTimestampMap removeObjectForKey:@([documentSnapshot hash])];
+  FIRServerTimestampBehavior serverTimestampBehavior =
+      [self toServerTimestampBehavior:timestampBehaviorString];
+
+  if (timestampBehaviorString != nil) {
+    [FLTFirebaseFirestorePlugin.serverTimestampMap removeObjectForKey:@([documentSnapshot hash])];
+  }
 
   return @{
     @"path" : documentSnapshot.reference.path,
@@ -191,11 +195,15 @@
   NSMutableArray *paths = [NSMutableArray array];
   NSMutableArray *documents = [NSMutableArray array];
   NSMutableArray *metadatas = [NSMutableArray array];
-  FIRServerTimestampBehavior serverTimestampBehavior =
-      [self toServerTimestampBehavior:FLTFirebaseFirestorePlugin
-                                          .serverTimestampMap[@([querySnapshot hash])]];
+  NSString *timestampBehaviorString =
+      FLTFirebaseFirestorePlugin.serverTimestampMap[@([querySnapshot hash])];
 
-  [FLTFirebaseFirestorePlugin.serverTimestampMap removeObjectForKey:@([querySnapshot hash])];
+  FIRServerTimestampBehavior serverTimestampBehavior =
+      [self toServerTimestampBehavior:timestampBehaviorString];
+
+  if (timestampBehaviorString != nil) {
+    [FLTFirebaseFirestorePlugin.serverTimestampMap removeObjectForKey:@([querySnapshot hash])];
+  }
 
   for (FIRDocumentSnapshot *document in querySnapshot.documents) {
     [paths addObject:document.reference.path];