+++ packages/cloud_firestore/cloud_firestore/lib/src/utils/codec_utility.dart
@@ -16,7 +16,30 @@ class _CodecUtility {
     return output;
   }
 
-  static List<dynamic>? replaceValueWithDelegatesInArray(List<dynamic>? data) {
+  static Map<FieldPath, dynamic>? replaceValueWithDelegatesInMapFieldPath(
+    Map<Object, dynamic>? data,
+  ) {
+    if (data == null) {
+      return null;
+    }
+    Map<FieldPath, dynamic> output = <FieldPath, dynamic>{};
+    data.forEach((key, value) {
+      if (key is FieldPath) {
+        output[key] = valueEncode(value);
+      } else if (key is String) {
+        output[FieldPath.fromString(key)] = valueEncode(value);
+      } else {
+        throw StateError(
+          'Invalid key type for map. Expected String or FieldPath, but got $key: ${key.runtimeType}.',
+        );
+      }
+    });
+    return output;
+  }
+
+  static List<dynamic>? replaceValueWithDelegatesInArray(
+    Iterable<dynamic>? data,
+  ) {
     if (data == null) {
       return null;
     }
@@ -50,7 +73,7 @@ class _CodecUtility {
   static dynamic valueEncode(dynamic value) {
     if (value is DocumentReference) {
       return value._delegate;
-    } else if (value is List) {
+    } else if (value is Iterable) {
       return replaceValueWithDelegatesInArray(value);
     } else if (value is Map<dynamic, dynamic>) {
       return replaceValueWithDelegatesInMap(value);