+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/utils/firestore_message_codec.dart
@@ -109,6 +109,8 @@ class FirestoreMessageCodec extends StandardMessageCodec {
     } else if (value is Settings) {
       buffer.putUint8(_kFirestoreSettings);
       writeValue(buffer, value.asMap);
+    } else if (value is Iterable && value is! List) {
+      super.writeValue(buffer, value.toList());
     } else if (value == double.nan) {
       buffer.putUint8(_kNaN);
     } else if (value == double.infinity) {