+++ packages/firebase_auth/firebase_auth_platform_interface/lib/src/id_token_result.dart
@@ -51,6 +51,6 @@ class IdTokenResult {
 
   @override
   String toString() {
-    return '$IdTokenResult(authTime: $authTime, claims: ${claims.toString()}, expirationTime: $expirationTime, issuedAtTime: $issuedAtTime, signInProvider: $signInProvider, token: $token)';
+    return '$IdTokenResult(authTime: $authTime, claims: $claims, expirationTime: $expirationTime, issuedAtTime: $issuedAtTime, signInProvider: $signInProvider, token: $token)';
   }
 }