+++ packages/firebase_ui_localizations/lib/src/lang/he.dart
@@ -1,0 +1,288 @@
+// Copyright 2022, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../default_localizations.dart';
+
+class HeLocalizations extends FirebaseUILocalizationLabels {
+  const HeLocalizations();
+
+  @override
+  String get accessDisabledErrorText =>
+      "הגישה לחשבון זה הושבתה באופן זמני"; //"Access to this account has been temporarily disabled"
+
+  @override
+  String get arrayLabel => "מערך"; //"array";
+
+  @override
+  String get booleanLabel => "בוליאני"; //"Boolean";
+
+  @override
+  String get cancelLabel => "בטל"; //"cancel";
+
+  @override
+  String get chooseACountry => "בחר מדינה"; //"Choose a country";
+
+  @override
+  String get confirmPasswordDoesNotMatchErrorText =>
+      "הסיסמאות אינן תואמות"; //"Passwords do not match";
+
+  @override
+  String get confirmPasswordInputLabel => "וודא סיסמה"; // "Confirm password";
+
+  @override
+  String get confirmPasswordIsRequiredErrorText =>
+      "וודא את הסיסמה שלך"; //"Confirm your password";
+
+  @override
+  String get continueText => "המשך"; //"Continue";
+
+  @override
+  String get countryCode => "קידומת"; //"Code";
+
+  @override
+  String get credentialAlreadyInUseErrorText =>
+      "ספק זה משויך לחשבון משתמש אחר."; //"This provider is associated with a different user account.";
+
+  @override
+  String get deleteAccount => "מחיקת חשבון"; //"Delete account";
+
+  @override
+  String get differentMethodsSignInTitleText =>
+      "השתמש באחת מהשיטות הבאות כדי להיכנס"; //"Use one of the following methods to sign in";
+
+  @override
+  String get disable => "השבת"; //"Disable";
+
+  @override
+  String get eastInitialLabel => "E";
+
+  @override
+  String get emailInputLabel => "דואר אלקטרוני"; //"Email";
+
+  @override
+  String get emailIsRequiredErrorText =>
+      "דואר אלקטרוני נדרש"; //"Email is required";
+
+  @override
+  String get emailLinkSignInButtonLabel =>
+      "כניסה באמצעות קישור קסם"; //"Sign in with magic link";
+
+  @override
+  String get emailTakenErrorText =>
+      """"חשבון עם דוא"ל זה כבר קיים"""; //"An account with this email already exists";
+
+  @override
+  String get enable => "הפעל"; //"Enable";
+
+  @override
+  String get enableMoreSignInMethods =>
+      "הפעל שיטות כניסה נוספות"; //"Enable more sign-in methods";
+
+  @override
+  String get enterSMSCodeText => "הזן קוד SMS"; //"Enter SMS code";
+
+  @override
+  String get findProviderForEmailTitleText =>
+      """הזן את הדוא"ל שלך כדי להמשיך"""; //"Enter your email to continue";
+
+  @override
+  String get forgotPasswordButtonLabel =>
+      "שכחת את הסיסמה?"; //"Forgotten password?";
+
+  @override
+  String get forgotPasswordHintText =>
+      """הזן את הדוא"ל שלך ואנו נשלח לך קישור לאיפוס הסיסמה"""; //"Provide your email and we will send you a link to reset your password";
+
+  @override
+  String get forgotPasswordViewTitle => "סיסמה נשכחה"; //"Forgotten password";
+
+  @override
+  String get geopointLabel => "גיאופוינט"; //"geopoint";
+
+  @override
+  String get goBackButtonLabel => "חזרה"; //"Go back";
+
+  @override
+  String get invalidCountryCode => "קוד מדינה לא חוקי"; //"Invalid code";
+
+  @override
+  String get isNotAValidEmailErrorText =>
+      """ספק דוא"ל חוקי"""; //Provide a valid email";
+
+  @override
+  String get latitudeLabel => "קו רוחב"; //"latitude";
+
+  @override
+  String get linkEmailButtonText => "הבא"; //"Next";
+
+  @override
+  String get longitudeLabel => "קו אורך"; //"longitude";
+
+  @override
+  String get mapLabel => "מפה"; //"map";
+
+  @override
+  String get mfaTitle => "אימות דו-שלבי"; //"2-Step Verification";
+
+  @override
+  String get name => "שם"; //"Name";
+
+  @override
+  String get northInitialLabel => "N";
+
+  @override
+  String get nullLabel => "null";
+
+  @override
+  String get numberLabel => "מספר"; //"number";
+
+  @override
+  String get off => "כבוי"; //"Off";
+
+  @override
+  String get on => "פעיל"; //"On";
+
+  @override
+  String get passwordInputLabel => "סיסמה"; //"Password";
+
+  @override
+  String get passwordIsRequiredErrorText =>
+      "נדרשת סיסמה"; //"Password is required";
+
+  @override
+  String get passwordResetEmailSentText =>
+      """שלחנו לך דוא"ל עם קישור לאיפוס הסיסמה שלך. אנא בדוק את הדוא"ל שלך."""; //"We've sent you an email with a link to reset your password. Please check your emails.";
+
+  @override
+  String get phoneInputLabel => "מספר טלפון"; //"Phone number";
+
+  @override
+  String get phoneNumberInvalidErrorText =>
+      "מספר הטלפון אינו חוקי"; //"Phone number is invalid";
+
+  @override
+  String get phoneNumberIsRequiredErrorText =>
+      "נדרש מספר טלפון"; //"Phone number is required";
+
+  @override
+  String get phoneVerificationViewTitleText =>
+      "הזן את מספר הטלפון שלך"; //"Enter your phone number";
+
+  @override
+  String get profile => "פרופיל"; //"Profile";
+
+  @override
+  String get provideEmail =>
+      """הזן את כתובת הדוא"ל והסיסמה שלך"""; //"Provide your email and password";
+
+  @override
+  String get referenceLabel => "הפניה"; //"reference";
+
+  @override
+  String get registerActionText => "הירשם"; //"Register";
+
+  @override
+  String get registerHintText => "אין לך חשבון?"; //"Don't have an account?";
+
+  @override
+  String get registerText => "הרשמה"; //"Register";
+
+  @override
+  String get resetPasswordButtonLabel => "איפוס סיסמה"; //"Reset password";
+
+  @override
+  String get sendLinkButtonLabel => "שלח קישור קסם"; //"Send magic link";
+
+  @override
+  String get signInActionText => "כנס"; //"Sign in";
+
+  @override
+  String get signInHintText => "כבר יש לך חשבון?"; //"Already have an account?";
+
+  @override
+  String get signInMethods => "שיטות כניסה"; //"Sign-in methods";
+
+  @override
+  String get signInText => "כניסה"; //"Sign in";
+
+  @override
+  String get signInWithAppleButtonText =>
+      "התחברות באמצעות חשבון Apple"; //"Sign in with Apple";
+
+  @override
+  String get signInWithEmailLinkSentText =>
+      """שלחנו לך דוא"ל עם קישור קסם. בדוק את הדואר האלקטרוני שלך ולחץ על הקישור כדי להיכנס"""; //"We've sent you an email with a magic link. Check your email and follow the link to sign in";
+
+  @override
+  String get signInWithEmailLinkViewTitleText =>
+      "כניסה באמצעות קישור קסם"; //"Sign in with magic link";
+
+  @override
+  String get signInWithFacebookButtonText =>
+      "כניסה באמצעות חשבון פייסבוק"; //"Sign in with Facebook";
+
+  @override
+  String get signInWithGoogleButtonText =>
+      "כניסה באמצעות חשבון של Google"; //"Sign in with Google";
+
+  @override
+  String get signInWithPhoneButtonText =>
+      "כניסה באמצעות מספר טלפון"; //"Sign in with phone";
+
+  @override
+  String get signInWithTwitterButtonText =>
+      "כניסה באמצעות חשבון טוויטר"; //"Sign in with Twitter";
+
+  @override
+  String get signOutButtonText => "התנתק"; //"Sign out";
+
+  @override
+  String get smsAutoresolutionFailedError =>
+      "ייבוא קוד ה-SMS באופן אוטומטי נכשל. אנא הזן את הקוד שלך באופן ידני"; //"Failed to resolve SMS code automatically. Please enter your code manually";
+
+  @override
+  String get southInitialLabel => "S";
+
+  @override
+  String get stringLabel => "מחרוזת"; //"string";
+
+  @override
+  String get timestampLabel => "חותמת זמן"; //"timestamp";
+
+  @override
+  String get typeLabel => "סוג"; //"type";
+
+  @override
+  String get unknownError =>
+      "אירעה שגיאה לא ידועה"; //"An unknown error occurred";
+
+  @override
+  String get updateLabel => "עדכון"; //"update";
+
+  @override
+  String get userNotFoundErrorText =>
+      "החשבון אינו קיים"; //"Account doesn't exist";
+
+  @override
+  String get valueLabel => "ערך"; //"value";
+
+  @override
+  String get verifyCodeButtonText => "אמת"; //"Verify";
+
+  @override
+  String get verifyingSMSCodeText => "מאמת קוד SMS..."; //"Verifying SMS code…";
+
+  @override
+  String get verifyItsYouText => "וודא שזה אתה"; //"Verify that it's you";
+
+  @override
+  String get verifyPhoneNumberButtonText => "הבא"; //"Next";
+
+  @override
+  String get westInitialLabel => "W";
+
+  @override
+  String get wrongOrNoPasswordErrorText =>
+      "הסיסמה אינה חוקית או שלמשתמש אין סיסמה"; //"The password is invalid or the user does not have a password";
+}