+++ packages/firebase_analytics/firebase_analytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 3.3.21
+
+ - Update a dependency to the latest release.
+
+## 3.3.20
+
+ - Update a dependency to the latest release.
+
+## 3.3.19
+
+ - Update a dependency to the latest release.
+
+## 3.3.18
+
+ - **DOCS**: add more documentation for `logEvent` ([#10389](https://github.com/firebase/flutterfire/issues/10389)). ([8755d776](https://github.com/firebase/flutterfire/commit/8755d776b30ce2db71e00279438c8d3617cf3187))
+
 ## 3.3.17
 
  - Update a dependency to the latest release.