+++ packages/firebase_app_check/firebase_app_check_platform_interface/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.5+15
+
+ - Update a dependency to the latest release.
+
+## 0.0.5+14
+
+ - Update a dependency to the latest release.
+
+## 0.0.5+13
+
+ - Update a dependency to the latest release.
+
+## 0.0.5+12
+
+ - Update a dependency to the latest release.
+
 ## 0.0.5+11
 
  - Update a dependency to the latest release.