+++ packages/firebase_in_app_messaging/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.7.0+14
+
+ - Update a dependency to the latest release.
+
+## 0.7.0+13
+
+ - Update a dependency to the latest release.
+
+## 0.7.0+12
+
+ - Update a dependency to the latest release.
+
+## 0.7.0+11
+
+ - Update a dependency to the latest release.
+
 ## 0.7.0+10
 
  - Update a dependency to the latest release.