+++ packages/firebase_messaging/firebase_messaging_platform_interface/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 4.2.14
+
+ - Update a dependency to the latest release.
+
+## 4.2.13
+
+ - Update a dependency to the latest release.
+
+## 4.2.12
+
+ - Update a dependency to the latest release.
+
+## 4.2.11
+
+ - Update a dependency to the latest release.
+
 ## 4.2.10
 
  - Update a dependency to the latest release.