+++ packages/firebase_ui_firestore/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.2.8
+
+ - Update a dependency to the latest release.
+
+## 1.2.7
+
+ - Update a dependency to the latest release.
+
+## 1.2.6
+
+ - Update a dependency to the latest release.
+
+## 1.2.5
+
+ - Update a dependency to the latest release.
+
+## 1.2.4
+
+ - **REFACTOR**: upgrade project to remove warnings from Flutter 3.7 ([#10344](https://github.com/firebase/flutterfire/issues/10344)). ([e0087c84](https://github.com/firebase/flutterfire/commit/e0087c845c7526c11a4241a26d39d4673b0ad29d))
+
 ## 1.2.3
 
  - Update a dependency to the latest release.